git/rev_list_scanner.go: read stderr in parallel
When processing the output of the "git rev-list" command, at present
we wait to read any data from the standard error pipe until after
reading all the data from the standard output pipe.  Specifically,
we attempt to read from the standard error pipe only in the goroutine
started by the Close() method of the RevListScanner structure in our
"git" package, and that method is not invoked by callers until they
have finished using the Scan() method to read and parse all the normal
output from the Git command.

However, the command may sometimes write enough data to its standard
error stream that it fills the kernel's buffer.  Because the stream is
opened to use blocking I/O, the subprocess is then prevented from making
further progress.  As we will not read from the standard error pipe
until we read EOF from the standard output pipe, and the subprocess
will not write to its standard output stream until we read from the
standard error pipe, both processes will stall indefinitely.

To resolve this issue we follow the same approach used in the
NewLsFiles() function of our "git" package and the parseScannerLogOutput()
function in our "lfs" package, whereby we start a separate goroutine
to read from the standard error pipe.  This ensures that the pipe
is drained as data is written to it and so its buffer cannot fill
to capacity.

Note that the Start() method of the Cmd structure in the "os/exec"
package of the standard Go library will by default use the same
technique for either or both of the standard output and error streams,
but only when the Stdout or Stderr elements of the Cmd structure,
respectively, are set to values that can not be cast as pointers to
the File structure from the "os" package.  In our case, though,
we use the StdoutPipe() and StderrPipe() methods of the Cmd structure,
which set those values to pointers to File structures, so we have
to implement our own goroutines to drain the pipes when necessary.