t/t-{checkout,pull}.sh: add read-only error tests
In a previous commit in this PR we adjusted the "git lfs checkout"
and "git lfs pull" commands so that when a Git LFS object's contents
can not be written to a file in the working tree, the commands output
a detailed set of error messages rather than just the message "could
not check out <filepath>".

We now add tests to our t-checkout.sh and t-pull.sh test scripts which
validate this change.  Specifically, the new "checkout: read-only
directory" and "pull: read-only directory" tests remove write permissions
on a directory in the working tree after removing a file which is
tracked as a Git LFS object from the directory.  The tests then check
that the "git lfs checkout" and "git lfs pull" commands output a full
set of error messages when they are unable to re-create the file.

Note that for historical reasons we expect that under these conditions
the commands will return a zero exit code, indicating success, rather
than a non-zero exit code, which would indicate failure.  We may adjust
this behaviour in the future, however, so that the commands also return
a non-zero exit code in these cases.