docs/man/git-lfs-config.adoc: use example domain
In the documentation of the "lfs.<url>.locksverify" configuration
option in our git-lfs-config(5) manual page, we use the github.com
domain name as an example, which may be slightly confusing to users,
as it could suggest the option is only available with GitHub's
services.

To better clarify the generic nature of this option, we update our
documentation to use the example.com domain reserved for such purposes.