docs/man/git-lfs-config.adoc: use literal block
In the documentation of the "lfs.<url>.locksverify" configuration
option in our git-lfs-config(5) manual page, we provide a somewhat
lengthy "git config" command example, which we currently format
as monospace text (but not literal monospace text) within a longer
paragraph.

To make this example "git config" command more readable, we adjust
its formatting so it is contained in a separate literal block.
This will ensure the example is rendered verbatim without automatic
text wrapping at the right margin, which will make it easier to parse
as a single command rather than two sequential ones.