docs/man/git-lfs-{config,faq}.adoc: variable names
In our git-lfs-config(5) and git-lfs-faq(7) manual pages we use several
different formats to express variable segments of Git configuration option
keys, including "<var>", "{var}", and "*".

In order to be as consistent as possible both internally and with Git's
own documentation, we now use the "<var>" format everywhere, and use the
same names for the variable segments as found in Git's documentation,
specifically "branch.<name>", "remote.<name>", "http.<url>", and
"url.<base>".  We also use "lfs.<url>" in cases where we expect to match
against a URL in the same fashion as for Git's "http.<url>.*" options.

In the section of our git-lfs-config(5) manual page which describes
the configuration options permitted in an ".lfsconfig" file, we
previously used curly braces to delimit the variable segment of
certain configuration option keys.  To avoid these segments being parsed
as references to AsciiDoc custom attributes, we also escaped the leading
curly brace.  We also used an asterisk for the "<url>" segment of the
"lfs.<url>.access" option rather than the term "url".

In practice, because AsciiDoc attributes must contain only ASCII
alphanumeric, underscore, and hyphen characters, the "{*}" segment was
not considered a possible valid attribute reference, and so the leading
backslash would be rendered as a literal character in the HTML and roff
versions of the manual page, e.g.:

  - lfs.\{*}.access
  - remote.{name}.lfsurl

By replacing the curly braces with angle brackets we avoid this problem
and can therefore also remove the leading backslashes.