docs/man/git-lfs-config.adoc: ext priority allowed
In the section of our git-lfs-config(5) manual page which describes
the configuration options permitted in an ".lfsconfig" file, we
enumerate those options in a list, which is intended to reflect the
options allowed by the readGitConfig() function in our "config" package.

Since commit ef3a44a03007ce35e217b0a8b0361b69994c6b87 of PR #760,
the set of options allowed in an ".lfsconfig" file has included
the "lfs.extension.<name>.priority" option, as is demonstrated by
the "extension config" test in our t/t-config.sh test script.

However, we do not at present include that option in the list in
our git-lfs-config(5) manual page, so we add it now.