docs/man/git-lfs-config.adoc: add section link
In our git-lfs-config(5) manual page the section on configuration files
contains a mention of the later section in the same page which describes
the configuration options permitted in an ".lfsconfig" file.  However,
this reference is not formatted as an AsciiDoc cross-reference, so it
is rendered as plain text.

We therefore revise this reference into an AsciiDoc cross-reference,
using the automatically generated anchor ID of the section which
describes the options allowed in an ".lfsconfig" file as the target
of the cross-reference.