docs/man: always use monospace format for literals
In our manual page AsciiDoc source files we are relatively consistent
in the use of monospace formatting for text that should be rendered
literally.

However, there are a number of exceptions where we use single quotes,
double quotes, or no delimiters at all, so we adjust those instances
now.  We make no other changes to the manual page text.