docs/man: use double quotes consistently
In several of our manual pages we quote several terms to indicate
they are meant to have a specific meaning unique to the implementation
of the Git LFS client.

For example, when discussing how the client determines if a cached
Git LFS object may be removed by the "git lfs prune" command, we use
the term "recent" with quotation marks to clarify that it is not meant
in a casual sense but is defined by the expiry times set with the
"lfs.fetchRecentCommitsDays", "lfs.fetchRecentRefsDays", and
"lfs.pruneOffsetDays" configuration options.

At present, we use a mix of single quotes and double quotes for these
types of terms.  Since we generally aim to follow the common formatting
conventions for North American English, we now replace any single quotes
with double quotes.

We also add quotation marks in several instances where a term was
inconsistently quoted, and in a few cases we add or remove words
in a quoted phrase to try to further clarify its intent.  As well,
we correct a minor typo in the git-lfs-migrate(1) manual page where
the word "from" was used instead of "form".