docs/man/git-lfs-track.adoc: fix option header
In commit f9fd1af3d5ae68fd2001f0cbf7f0c79468371557 of PR #5054 we
corrected the indentation in some of our manual page source files,
following their conversion to the AsciiDoc format in commit
0c66dcf15d988802255c3133cd5ab8105cbf0ef7 in the same PR.  For a number
of pages these corrections entailed reformatting the lists of options
supported by the different commands into description lists, where
each entry consists of one or more terms, followed by a double-colon
delimiter, followed by a description of the term or terms.

In the case of the git-lfs-track(1) manual page, though, the second
colon character was accidentally omitted from the term of the
final entry in the list of command options.  Specifically, because
the entry for the --no-modify-attrs option lacks a full term delimiter,
it does not render properly when converted to HTML or groff, so we
correct this problem now.