docs/man/git-lfs-{config,migrate}.adoc: tidy lists
Our git-lfs-config(5) and git-lfs-migrate(1) manual pages contain
several second-level lists which enumerate the values accepted for
specific configuration and command options.

While the entries in these second-level lists are relatively consistent,
they are not formatted identically, with some using a hyphen to
separate the initial literal terms from the following description,
and others using a colon or nothing at all.  As well, some descriptions
end with a period, and others do not.

In subsequent commits in this PR we will update the formatting of
these second-level lists so they render properly when converted to
HTML or roff by Asciidoctor, and also revise our "mangen" program so it
supports second-level lists in the plain-text versions of our manual
pages that it generates.

Before making these changes, we first update the text of the entries
in the second-level lists so it as consistent as possible both
within and between the lists.