docs/man/git-lfs-{config,migrate}.adoc: fix lists
Our git-lfs-config(5) and git-lfs-migrate(1) manual pages contain
several second-level lists which enumerate the values accepted for
specific configuration and command options.

Our second-level lists appear within top-level list entries that
contain one or more paragraphs attached to the principal text with
list continuations, which are lines with a single "+" character.
As such, although we use a distinct marker for our second-level list
entries (namely two asterisks instead of just one), this is not
actually sufficient, as the AsciiDoc specification requires that
we use open blocks to identify all the elements which should belong
to the parent top-level list entries.  See, for reference:

  https://docs.asciidoctor.org/asciidoc/latest/lists/continuation/#list-continuation

We could reformat these top-level entries to use a single list
continuation, followed by an open block which contains the entire
contents of the entry, including our nested lists along with all
the separate paragraphs of text.  This would be relatively elegant,
but would require more extensive changes in our "mangen" program
to recognize this alternative formatting technique for multi-paragraph
list entries.  At the moment, when parsing lists, our program simply
discards lines with a single "+" character, and expects that an
empty list signals the end of list.

The approach we take instead is to enclose just our second-level
lists in open blocks with the appropriate double-hyphen delimiters
on separate lines before and after the lists.  To attach these open
blocks to the preceding paragraphs in their top-level list entries
we use a list continuation line (i.e., a single "+" character).

In two cases, we also need to attach a paragraph which follows the
open blocks with the nested lists to the surrounding top-level list
entries.  To do this, we retain the existing list continuation lines
but convert them into ancestor list continuations by adding an empty
line before the line with the "+" character.  See, for reference:

  https://docs.asciidoctor.org/asciidoc/latest/lists/continuation/#ancestor-list-continuation

This change ensures that the paragraphs following our second-level
lists are not also indented as far as the entries of those nested lists,
at least when we render them with Asciidoctor as HTML or roff.  In a
subsequent commit in this PR we will revise our "mangen" program to
support these ancestor list continuations as well, so the plain-text
versions it generates of our manual pages will also contain correctly
formatted nested lists.