docs/man: use consistent Git and LFS program names
In several instances in our manual pages we refer to the Git LFS client
program as "git-lfs", which is technically accurate as that is the
name of the compiled binary program (except on Windows, where it is
"git-lfs.exe").

However, this is somewhat unintuitive for most users, who invoke the
Git LFS client through Git by running commands such as "git lfs fetch".

Therefore, we revise our manual pages so they refer to the Git LFS
client as such, rather than by the name of its binary.

We also correct one instance where we use the term "GIT LFS", with
"GIT" fully capitalized, to read "Git LFS" instead, and correct
another instance where we use the term "lfs" rather than "LFS".

As well, we ensure that we refer to "Git" rather than "git" when
we are not providing example commands.