docs/man/git-lfs-config.adoc: add lfs.gitProtocol
In PR #964 the "lfs.gitProtocol" configuration option was introduced
as a way to customize the URL scheme to be used in place of the "git"
scheme in the Git LFS server API URL.  By default, the "git" scheme
will be replaced with the "https" scheme, but this option allows for
another choice such as "http".

Although the configuration option is listed in our git-lfs-config(5)
manual page as one of the options that is allowed in ".lfsconfig" files,
it is otherwise undocumented, so we add an entry for it now in that
same manual page.