docs/man: fix option description indentation
In commit f9fd1af3d5ae68fd2001f0cbf7f0c79468371557 of PR #5054 we
corrected the indentation in some of our manual page source files,
following their conversion to the AsciiDoc format in commit
0c66dcf15d988802255c3133cd5ab8105cbf0ef7 in the same PR.  For a number
of pages these corrections entailed reformatting the lists of options
supported by the different commands into description lists, where
each entry consists of one or more terms, followed by a double-colon
delimiter, followed by a description of the term or terms.

In the case of the git-lfs-migrate(1) manual page, the description
text of the entry for the --skip-fetch option to the "git lfs migrate"
command was fully indented except for the final line, likely because it
begins with a monospace delimiter, so we adjust that now.

Similarly, the second line of the description of the --theirs option
in the git-lfs-checkout(1) manual page is not indented to match
the first line, so we add the indentation now.