docs/man/git-lfs-migrate.adoc: fix parameter args
In our git-lfs-migrate(1) manual page we describe a number of options
to the "git lfs migrate" command which accept a parameter argument, and
following our normal convention and that of Git's own documentation,
we generally use angle brackets around the name of the parameter to
indicate that the user should substitute their own value for the argument.

However, in several places in the manual page we refer to these
parameters but do not include the angle brackets around their names,
so we revise those instances now to match our usual formatting style
for parameter names.

Note that we always enclose these parameter names in AsciiDoc monospace
delimiter as well.