docs/man: escape links and bold within monospace
In our manual page AsciiDoc source files we are relatively consistent
in the use of monospace formatting for text that should be rendered
literally.

AsciiDoc monospace delimiters, however, do not prevent the interpretation
of other formatting marks within the enclosed text.  To achieve that,
we would need to use monospace literal delimiters, i.e., "{+ ... +}",
but our "mangen" program is not able to parse those at the moment, so
we continue to use simple monospace formatting for literals as it
suffices in the large majority of cases.

There are a few instances in our manual pages, though, where certain
elements within monospace text are interpreted as AsciiDoc formatting
marks when we do not intend for this to be the case, leading to problems
when we use Asciidoctor to render our source files into Ronn or HTML.

In our git-lfs-config(5) and git-lfs-faq(7) pages we include HTTP URLs
within monospace delimiters, and because they are preceded by a space
character, they are recognized by Asciidoctor as links and rendered
accordingly.  The consequence, particular in our Ronn output files,
is that the following paragraphs are underlined when they should not be.

To avoid this problem, we add a backslash character in front of the
HTTP URLs, which disables the automatic recognition of links in AsciiDoc.

For a similar reason, we also escape the leading asterisk characters in
examples of wildcard file patterns that appear in our git-lfs-fetch(1)
and git-lfs-faq(7) pages.  By adding a backslash before the asterisk,
we avoid Asciidoctor rendering the following sentence as bold text,
as it does now.

Note that we already use a backslash escape character before another
asterisk within monospaced text elsewhere in the git-lfs-faq(7) page,
in the entry regarding large file errors from GitHub, which prevents
the following paragraph being rendered as bold text.

Also note that our "mangen" program does not yet support these
escape characters within monospace text, so in a subsequent commit
in this PR we will modify the program to detect them and ignore them.