docs/man: add all missing short options
Many Git LFS client commands provide both short and long versions of
certain options, such as both -d and --dry-run to run a command in a
read-only mode, or both -f and --force to require that a command
proceed despite certain conditions.

In our manual pages, when we enumerate the list of available options
for a given command, if an option has a short version as well as long
one we usually list the short version first.  This is in keeping with
the practices followed by Git's own documentation.

However, in some of our manual pages we do not list the short version
of certain options, only the long version, so we add all the missing
short versions now.