docs/man/git-lfs-[un]install.adoc: add file option
The "git lfs install" and "git lfs uninstall" commands both accept
a --file option, which was introduced in commit
84ca7e7875d1a32c07dc56c8ad0e33f38f59a824 of PR #5355, which cause
the commands to add or remove Git LFS filter settings from a specified
Git configuration file rather than from any of the system, global,
local, or current working tree configuration files.

However, this option is not documented in either of our
git-lfs-install(1) or git-lfs-uninstall(1) manual pages, so we now
add entries for it into the lists of options supported by those commands.