docs/man/git-lfs-clone.adoc: deprecate command
The "git lfs clone" command has been deprecated since Git LFS v2.3.0
and Git v2.15.0 were released, so we update our manual page for the
command to reflect this fact and clarify when and why the command
should still be used, if at all.

As discussed in #2466, Git's support for long-running filter processes
(introduced in Git v2.11.0 via the "filter.<driver>.process" setting) was
enhanced in v2.15.0 with the "delay" capability whereby the filter can
signal that it will operate asynchronously on multiple blobs, thus allowing
it to process multiple blobs in parallel.  Git LFS was updated to make
use of the "delay" capability in a series of PRs culminating in PR #2511,
all of which were included in the Git LFS v2.3.0 release.