docs/man: use consistent parameter assignments
A number of Git LFS commands accept options which require a parameter
to be supplied as an argument after the option name, such as for
the --include and --exclude options supported by the "git lfs fetch"
abd "git lfs pull" commands.

In our manual pages, when we document these options and their required
parameters, we usually separate the long version of the option name
and its argument with an equals sign, but separate the short version of
the option name and its argument with a space character, if the option
has a short version at all.  This is in keeping with the practices
followed by Git's own documentation.

However, in some of our manual pages we use a space character to
separate the long version of an option from its argument.  While Git LFS
commands will accept arguments specified in this manner, in order to
make our manual pages as consistent as possible both with each other
and with Git's documentation, we revise our git-lfs-checkout(1) and
git-lfs-merge-driver(1) manual pages so they also use an equals sign as
a separator in the descriptions of all their options that require
parameters.