docs/man/git-lfs-pointer.adoc: add parameters
Since the introduction of the "git lfs pointer" command in PR #246
it has accepted --file and --pointer options, each of which require
a file path as a parameter.

However, our git-lfs-pointer(1) manual page does not include these
required parameters in the description list entries for the --file
and --pointer options, so we add them now.

We also take the opportunity to revise the command prototypes in
the synopsis section to refer to the same "<file>" parameter name
we use in the option description list entries, which will make the
terminology of our git-lfs-pointer(1) manual page more internally
consistent as well as more consistent with our other manual pages
and with the Git project's documentation.