docs/man/git-lfs-prune.adoc: edit verify parameter
In commit 2be5d269bb787e96302b64655f09839f423e1afd of PR #5648 we
enhanced the "git lfs prune" command, in part by adding support for a
new --when-unverified option which controls how the command behaves
when the --verify-remote option is specified and an object which
would otherwise be pruned is not verified as present on the current
Git remote.

In our git-lfs-prune(1) manual page we describe the --when-unverified
option as accepting a parameter whose name we give as "<halt,continue>",
with the intent that this should indicate that only the values "halt"
and "continue" will be accepted for this option.

However, following both our own convention and that of Git's own
documentation, we generally use angle brackets around the name of a
parameter to indicate that the user should substitute their own value
for the argument, and then refer to that parameter by that name, with
the angle brackets included.

So as to make the git-lfs-prune(1) manual page more consistent with
this practice, we change the name of the --when-unverified option's
parameter to "<action>", and then expand the description of the option
to explain the two acceptable values for that parameter, namely "halt"
(the default) and "continue".