docs/man/git-lfs-merge-driver.adoc: required args
In commit 402e958bfa121818dabb0aba2ff7b6865dfe9dbf of PR #4970 we
implemented the new "git lfs merge-driver" command, which can be
configured as a custom Git merge driver in order to help automatically
merge some types of files when they are tracked as Git LFS objects.

We expect this command to be invoked by Git during a merge operation,
and so several of its command-line flags are required, as their arguments
must specify the paths of the temporary files created by Git for use
by a merge driver program.

However, our git-lfs-merge-driver(1) manual page lists these required
flags and their file path parameters as options, without any indication
that they are actually required.  As well, one of the required flags,
the --output flag, is not described at all.

We therefore revise this manual page to document the mandatory flags,
including the --output flag.  We also take the opportunity to expand
the description of these flags in order to provide somewhat more
context as to their meaning, and we adjust a few other minor details
in the page as well.