Makefile,docs/man: undo extra synopsis roff indent
In a previous commit in this PR we revised all the synopsis sections
of our AsciiDoc manual pages so they were formatted consistently using
source listing blocks, which helps us avoid problems with missing
hard line breaks, among other advantages.

However, this change introduces one small but noticeable inconsistency
in the way our manual pages are now displayed by the man(1) program,
due to how the Asciidoctor "manpage" converter handles listing blocks
when generating roff output.  The "manpage" converter always indents
listing and literal blocks by a fixed amount, which is currently set
to four space characters, as seen in the convert_listing() and
convert_literal() methods of the Converter::ManPageConverter Ruby class:

  https://github.com/asciidoctor/asciidoctor/blob/99bc6105e021bf68d8c1018959d6d5d91ffcb2f3/lib/asciidoctor/converter/manpage.rb#L253-L260
  https://github.com/asciidoctor/asciidoctor/blob/99bc6105e021bf68d8c1018959d6d5d91ffcb2f3/lib/asciidoctor/converter/manpage.rb#L269-L276

These methods enclose the content of the blocks with a set of roff
macros, two of which in particular indent and unindent the text.
In general, this is desirable, as we expect to use source listing and
literal blocks within a series of regular paragraphs, and so rendering
them in an indented fashion makes them distinct and recognizable as
examples of source code or other literal text.

Our synopsis sections, though, explicitly do not contain any content
other than Git LFS command prototypes, which we would like to align
with the main content of the other sections, both for visual consistency
and because this is the standard format used for all Unix manual pages.

Because the roff indentation macros are, at present, hard-coded into
the Asciidoctor "manpage" converter's methods, we can not override
them with simple AsciiDoc options or flags.

Instead, to achieve the formatting we desire for our synopsis sections,
we introduce a small Asciidoctor extension module which subclasses
the Converter::ManPageConverter class and provides its own definition
of the convert_literal() method.  In this method, we first call the
parent class's method of the same name, and then modify its output
to remove the roff indentation and unindentation macros, but only
if the current document node's "role" attribute is set to "synopsis".

To locate the indentation and unindentation macros in our extension
module, we define a regular expression which matches the roff macros
output by the convert_literal() method of the Converter::ManPageConverter
class, which are the ".RS" ("region start") and ".RE" ("region end")
indentation macros.  We expect the former will be followed by an
numeric value, which currently is always "4", but we allow for any
possible value here in case Asciidoctor changes the indentation amount
for listing blocks in the future.

We also allow for these macros to be preceded by ".if" conditional
requests with the "n" condition, which is how Asciidoctor generates
them at present.  The "n" (i.e., "nroff") condition is normally set
true by default if the output device is a terminal.  See, for reference:

  https://www.gnu.org/software/groff/manual/groff.html#Indented-regions-in-ms
  https://www.gnu.org/software/groff/manual/groff.html#troff-and-nroff-Modes

Finally, we add a custom "role" attribute to the source listing blocks
in all of our manual pages' synopsis sections, and we update our Makefile
so Asciidoctor will find and load our new extension module when generating
roff output with the "manpage" converter.