docs/man/git-lfs-pre-push.adoc: drop ignored arg
In commit 847a547f3468d18347c8d82bd3da65efcd4cc1bf of PR #263
we refactored our original version of the "git lfs push" command
and created a new "git lfs pre-push" command, which we register
as a Git hook.

We also added a git-lfs-pre-push(1) manual page for this command,
in commit 9fe3a5b21274e275b1a6489f44d55c0f87a88996 of the same PR.
Since that time, this page has always indicated that the command
accepts an initial "<remote>" argument followed by an optional
"remoteurl" argument.

This description is somewhat accurate, in that Git invokes pre-push
hook commands with two parameters.  The first is the name of the
push operation's destination remote, and the second is the location
of that remote.  If the user did not specify a named remote, then
both arguments are identical and contain the remote's location.

Our "git lfs pre-push" command simply disregards the second parameter
entirely, however, rather than treating it as an optional argument
which will be read if it is provided.  We therefore adjust our
manual page now to clarify that the command requires just a single
argument, and ignores the second one provided by Git.